refactor: ProgressItem 파일 병합 및 DOM 최적화
